% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpapp.R
\name{hpca_app}
\alias{hpca_app}
\title{interactive use of HumanPrimaryCellAtlasData}
\usage{
hpca_app(...)
}
\arguments{
\item{\dots}{passed to `do_SingleR`}
}
\description{
interactive use of HumanPrimaryCellAtlasData
}
\examples{
if (interactive()) hpca_app()
}
