% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_SingleR.R
\name{ingest_h5ad}
\alias{ingest_h5ad}
\title{ingestion and conversion}
\usage{
ingest_h5ad(path = "/home/vincent/tenx3k.h5ad")
}
\arguments{
\item{path}{character(1) relative path to h5ad file}
}
\description{
ingestion and conversion
}
